Move skill to time_based and remove all old models
